:wrench: chore: initial organization
