+++ src/sentry/notifications/notification_action/__init__.py
@@ -1,0 +1,8 @@
+__all__ = ["group_type_notification_registry", "issue_alert_handler_registry"]
+
+from sentry.utils.registry import Registry
+
+from .issue_alert_registry import issue_alert_handler_registry
+from .types import LegacyRegistryHandler
+
+group_type_notification_registry = Registry[LegacyRegistryHandler]()
