+++ src/sentry/notifications/notification_action/exceptions.py
@@ -1,0 +1,2 @@
+class NotificationHandlerException(Exception):
+    pass
