+++ src/sentry/notifications/notification_action/issue_alert_registry/handlers/azure_devops_issue_alert_handler.py
@@ -1,0 +1,12 @@
+from sentry.notifications.notification_action.issue_alert_registry import (
+    issue_alert_handler_registry,
+)
+from sentry.notifications.notification_action.issue_alert_registry.base import (
+    TicketingIssueAlertHandler,
+)
+from sentry.workflow_engine.models import Action
+
+
+@issue_alert_handler_registry.register(Action.Type.AZURE_DEVOPS)
+class AzureDevopsIssueAlertHandler(TicketingIssueAlertHandler):
+    pass
