+++ src/sentry/workflow_engine/handlers/__init__.py
@@ -9,7 +9,7 @@ __all__ = [
     "GithubEnterpriseActionHandler",
     "JiraActionHandler",
     "JiraServerActionHandler",
-    "MsteamsActionHandler",
+    "MSTeamsActionHandler",
     "OpsgenieActionHandler",
     "PagerdutyActionHandler",
     "PluginActionHandler",
@@ -26,7 +26,7 @@ from .action import (
     GithubEnterpriseActionHandler,
     JiraActionHandler,
     JiraServerActionHandler,
-    MsteamsActionHandler,
+    MSTeamsActionHandler,
     OpsgenieActionHandler,
     PagerdutyActionHandler,
     PluginActionHandler,
