+++ src/sentry/workflow_engine/handlers/action/notification/base.py
@@ -1,0 +1,77 @@
+import logging
+from abc import ABC, abstractmethod
+
+from sentry.notifications.models.notificationaction import ActionTarget
+from sentry.notifications.notification_action.utils import execute_via_issue_alert_handler
+from sentry.workflow_engine.models import Action, Detector
+from sentry.workflow_engine.types import ActionHandler, WorkflowEventData
+
+logger = logging.getLogger(__name__)
+
+
+class LegacyRegistryInvoker(ABC):
+    """
+    Abstract base class that defines the interface for notification handlers.
+    """
+
+    @staticmethod
+    @abstractmethod
+    def handle_workflow_action(job: WorkflowEventData, action: Action, detector: Detector) -> None:
+        """
+        Implement this method to handle the specific notification logic for your handler.
+        """
+        raise NotImplementedError
+
+
+class IntegrationActionHandler(ActionHandler, ABC):
+    # TODO(iamrajjoshi): Switch this to an enum after we decide on what this enum will be
+    provider_slug: str
+
+
+class TicketingActionHandler(IntegrationActionHandler, ABC):
+    config_schema = {
+        "$schema": "https://json-schema.org/draft/2020-12/schema",
+        "description": "The configuration schema for a Ticketing Action",
+        "type": "object",
+        "properties": {
+            "target_identifier": {
+                "type": ["null"],
+            },
+            "target_display": {
+                "type": ["null"],
+            },
+            "target_type": {
+                "type": ["integer"],
+                "enum": [ActionTarget.SPECIFIC.value],
+            },
+        },
+    }
+
+    data_schema = {
+        "$schema": "https://json-schema.org/draft/2020-12/schema",
+        "type": "object",
+        "description": "Schema for ticket creation action data blob",
+        "properties": {
+            "dynamic_form_fields": {
+                "type": "array",
+                "description": "Dynamic form fields from customer configuration",
+                "items": {"type": "object"},
+                "default": [],
+            },
+            "additional_fields": {
+                "type": "object",
+                "description": "Additional fields that aren't part of standard fields",
+                "additionalProperties": True,
+                "default": {},
+            },
+        },
+        "additionalProperties": False,
+    }
+
+    @staticmethod
+    def execute(
+        job: WorkflowEventData,
+        action: Action,
+        detector: Detector,
+    ) -> None:
+        execute_via_issue_alert_handler(job, action, detector)
