+++ src/sentry/workflow_engine/handlers/action/notification/common.py
@@ -1,0 +1,65 @@
+from sentry.notifications.models.notificationaction import ActionTarget
+
+# TODO(iamrajjoshi): This should be removed once I define the config schemas for each action type
+GENERIC_ACTION_CONFIG_SCHEMA = {
+    "$schema": "https://json-schema.org/draft/2020-12/schema",
+    "description": "The configuration schema for Notification Actions",
+    "type": "object",
+    "properties": {
+        "target_identifier": {
+            "type": ["string", "null"],
+        },
+        "target_display": {
+            "type": ["string", "null"],
+        },
+        "target_type": {
+            "type": ["integer", "null"],
+            "enum": [*ActionTarget] + [None],
+        },
+    },
+}
+
+MESSAGING_ACTION_CONFIG_SCHEMA = {
+    "$schema": "https://json-schema.org/draft/2020-12/schema",
+    "description": "The configuration schema for a Messaging Action",
+    "type": "object",
+    "properties": {
+        "target_identifier": {"type": ["string"]},
+        "target_display": {"type": ["string"]},
+        "target_type": {
+            "type": ["integer"],
+            "enum": [ActionTarget.SPECIFIC.value],
+        },
+    },
+    "required": ["target_identifier", "target_display", "target_type"],
+    "additionalProperties": False,
+}
+
+TAGS_SCHEMA = (
+    {
+        "type": "string",
+        "description": "Tags to add to the message",
+    },
+)
+
+NOTES_SCHEMA = {
+    "type": "string",
+    "description": "Notes to add to the message",
+}
+
+
+ONCALL_ACTION_CONFIG_SCHEMA = {
+    "$schema": "https://json-schema.org/draft/2020-12/schema",
+    "description": "The configuration schema for a on-call Action",
+    "type": "object",
+    "properties": {
+        "target_identifier": {"type": ["string"]},
+        "target_display": {"type": ["string", "null"]},
+        "target_type": {
+            "type": ["integer"],
+            "enum": [ActionTarget.SPECIFIC.value],
+        },
+    },
+    "required": ["target_identifier", "target_type"],
+    "additionalProperties": False,
+}
